+++ crates/gqlitedb/src/interpreter/expression_analyser.rs
@@ -36,6 +36,26 @@ pub(crate) trait Variables
   fn expression_type(&self, name: impl Into<String>) -> Result<ExpressionType>;
 }
 
+mod validators
+{
+  use crate::{error, Result};
+
+  use super::{ExpressionInfo, ExpressionType};
+
+  pub(super) fn any(x: ExpressionInfo) -> Result<ExpressionInfo>
+  {
+    Ok(x)
+  }
+  pub(super) fn boolean_or_null(x: ExpressionInfo) -> Result<ExpressionInfo>
+  {
+    match x.expression_type
+    {
+      ExpressionType::Boolean | ExpressionType::Null | ExpressionType::Variant => Ok(x),
+      _ => Err(error::CompileTimeError::InvalidArgumentType.into()),
+    }
+  }
+}
+
 impl ExpressionInfo
 {
   fn new(expression_type: ExpressionType, constant: bool, aggregation_result: bool) -> Self
@@ -59,12 +79,14 @@ impl ExpressionInfo
     variables: &impl Variables,
     function_manager: &functions::Manager,
     expressions: impl Iterator<Item = &'a ast::Expression>,
+    validator: impl Fn(ExpressionInfo) -> Result<ExpressionInfo>,
   ) -> Result<Vec<ExpressionInfo>>
   {
     expressions
-      .map(|x| Self::analyse(variables, function_manager, x))
+      .map(|x| validator(Self::analyse(variables, function_manager, x)?))
       .collect::<Result<_>>()
   }
+
   pub(crate) fn analyse(
     variables: &impl Variables,
     function_manager: &functions::Manager,
@@ -75,11 +97,21 @@ impl ExpressionInfo
     {
       ast::Expression::Array(arr) => Ok(Self::new_type(
         ExpressionType::Array,
-        Self::analyses(variables, function_manager, arr.array.iter())?,
+        Self::analyses(
+          variables,
+          function_manager,
+          arr.array.iter(),
+          validators::any,
+        )?,
       )),
       ast::Expression::FunctionCall(call) =>
       {
-        let arguments = Self::analyses(variables, function_manager, call.arguments.iter())?;
+        let arguments = Self::analyses(
+          variables,
+          function_manager,
+          call.arguments.iter(),
+          validators::any,
+        )?;
         Ok(Self::new(
           function_manager.validate_arguments(
             &call.name,
@@ -92,19 +124,39 @@ impl ExpressionInfo
       }
       ast::Expression::IsNull(isn) => Ok(Self::new_type(
         ExpressionType::Boolean,
-        Self::analyses(variables, function_manager, [&isn.value].into_iter())?,
+        Self::analyses(
+          variables,
+          function_manager,
+          [&isn.value].into_iter(),
+          validators::any,
+        )?,
       )),
       ast::Expression::IsNotNull(isn) => Ok(Self::new_type(
         ExpressionType::Boolean,
-        Self::analyses(variables, function_manager, [&isn.value].into_iter())?,
+        Self::analyses(
+          variables,
+          function_manager,
+          [&isn.value].into_iter(),
+          validators::any,
+        )?,
       )),
       ast::Expression::Negation(ln) => Ok(Self::new_type(
         ExpressionType::Variant,
-        Self::analyses(variables, function_manager, [&ln.value].into_iter())?,
+        Self::analyses(
+          variables,
+          function_manager,
+          [&ln.value].into_iter(),
+          validators::any,
+        )?,
       )),
       ast::Expression::LogicalNegation(ln) => Ok(Self::new_type(
         ExpressionType::Boolean,
-        Self::analyses(variables, function_manager, [&ln.value].into_iter())?,
+        Self::analyses(
+          variables,
+          function_manager,
+          [&ln.value].into_iter(),
+          validators::boolean_or_null,
+        )?,
       )),
       ast::Expression::LogicalAnd(rd) => Ok(Self::new_type(
         ExpressionType::Boolean,
@@ -112,6 +164,7 @@ impl ExpressionInfo
           variables,
           function_manager,
           [&rd.left, &rd.right].into_iter(),
+          validators::boolean_or_null,
         )?,
       )),
       ast::Expression::LogicalOr(rd) => Ok(Self::new_type(
@@ -120,6 +173,7 @@ impl ExpressionInfo
           variables,
           function_manager,
           [&rd.left, &rd.right].into_iter(),
+          validators::boolean_or_null,
         )?,
       )),
       ast::Expression::LogicalXor(rd) => Ok(Self::new_type(
@@ -128,6 +182,7 @@ impl ExpressionInfo
           variables,
           function_manager,
           [&rd.left, &rd.right].into_iter(),
+          validators::boolean_or_null,
         )?,
       )),
       ast::Expression::RelationalEqual(rd) => Ok(Self::new_type(
@@ -136,6 +191,7 @@ impl ExpressionInfo
           variables,
           function_manager,
           [&rd.left, &rd.right].into_iter(),
+          validators::any,
         )?,
       )),
       ast::Expression::RelationalDifferent(rd) => Ok(Self::new_type(
@@ -144,6 +200,7 @@ impl ExpressionInfo
           variables,
           function_manager,
           [&rd.left, &rd.right].into_iter(),
+          validators::any,
         )?,
       )),
       ast::Expression::RelationalInferior(rd) => Ok(Self::new_type(
@@ -152,6 +209,7 @@ impl ExpressionInfo
           variables,
           function_manager,
           [&rd.left, &rd.right].into_iter(),
+          validators::any,
         )?,
       )),
       ast::Expression::RelationalSuperior(rd) => Ok(Self::new_type(
@@ -160,6 +218,7 @@ impl ExpressionInfo
           variables,
           function_manager,
           [&rd.left, &rd.right].into_iter(),
+          validators::any,
         )?,
       )),
       ast::Expression::RelationalInferiorEqual(rd) => Ok(Self::new_type(
@@ -168,6 +227,7 @@ impl ExpressionInfo
           variables,
           function_manager,
           [&rd.left, &rd.right].into_iter(),
+          validators::any,
         )?,
       )),
       ast::Expression::RelationalSuperiorEqual(rd) => Ok(Self::new_type(
@@ -176,6 +236,7 @@ impl ExpressionInfo
           variables,
           function_manager,
           [&rd.left, &rd.right].into_iter(),
+          validators::any,
         )?,
       )),
       ast::Expression::RelationalIn(ri) => Ok(Self::new_type(
@@ -184,6 +245,7 @@ impl ExpressionInfo
           variables,
           function_manager,
           [&ri.left, &ri.right].into_iter(),
+          validators::any,
         )?,
       )),
       ast::Expression::RelationalNotIn(ri) => Ok(Self::new_type(
@@ -192,6 +254,7 @@ impl ExpressionInfo
           variables,
           function_manager,
           [&ri.left, &ri.right].into_iter(),
+          validators::any,
         )?,
       )),
       ast::Expression::Addition(ri) => Ok(Self::new_type(
@@ -200,6 +263,7 @@ impl ExpressionInfo
           variables,
           function_manager,
           [&ri.left, &ri.right].into_iter(),
+          validators::any,
         )?,
       )),
       ast::Expression::Multiplication(ri) => Ok(Self::new_type(
@@ -208,6 +272,7 @@ impl ExpressionInfo
           variables,
           function_manager,
           [&ri.left, &ri.right].into_iter(),
+          validators::any,
         )?,
       )),
       ast::Expression::Substraction(ri) => Ok(Self::new_type(
@@ -216,6 +281,7 @@ impl ExpressionInfo
           variables,
           function_manager,
           [&ri.left, &ri.right].into_iter(),
+          validators::any,
         )?,
       )),
       ast::Expression::Division(ri) => Ok(Self::new_type(
@@ -224,6 +290,7 @@ impl ExpressionInfo
           variables,
           function_manager,
           [&ri.left, &ri.right].into_iter(),
+          validators::any,
         )?,
       )),
       ast::Expression::Modulo(ri) => Ok(Self::new_type(
@@ -232,11 +299,17 @@ impl ExpressionInfo
           variables,
           function_manager,
           [&ri.left, &ri.right].into_iter(),
+          validators::any,
         )?,
       )),
       ast::Expression::Map(map) => Ok(Self::new_type(
         ExpressionType::Map,
-        Self::analyses(variables, function_manager, map.map.values())?,
+        Self::analyses(
+          variables,
+          function_manager,
+          map.map.values(),
+          validators::any,
+        )?,
       )),
       ast::Expression::MemberAccess(ma) => Ok(Self::new_type(
         ExpressionType::Variant,